+++ apps/assets/api/__init__.py
@@ -1,11 +1,12 @@
-from .mixin import *
-from .category import *
-from .platform import *
-from .asset import *
-from .label import *
 from .account import *
-from .node import *
-from .domain import *
+from .asset import *
 from .automations import *
-from .gathered_user import *
+from .category import *
+from .domain import *
 from .favorite_asset import *
+from .gathered_user import *
+from .label import *
+from .mixin import *
+from .node import *
+from .platform import *
+from .tree import *
